Laravel opdracht 15: rolgebaseerde toegang
  ✓ student can access projects.index                                                                                                                   0.28s
  ✓ student can access projects.create                                                                                                                  0.05s
  ✓ student can store a new project                                                                                                                     0.05s
  ✓ student can access projects.edit                                                                                                                    0.05s
  ✓ student can update a project                                                                                                                        0.05s
  ✓ student cannot access projects.delete                                                                                                               0.05s
  ✓ student cannot destroy a project                                                                                                                    0.05s
  ✓ teacher can access projects.index                                                                                                                   0.05s
  ✓ teacher can access projects.create                                                                                                                  0.05s
  ✓ teacher can store a new project                                                                                                                     0.05s
  ✓ teacher can access projects.edit                                                                                                                    0.05s
  ✓ teacher can update a project                                                                                                                        0.05s
  ✓ teacher can access projects.delete                                                                                                                  0.05s
  ✓ teacher can destroy a project                                                                                                                       0.05s
  ✓ admin can access projects.index                                                                                                                     0.05s
  ✓ admin can access projects.create                                                                                                                    0.05s
  ✓ admin can store a new project                                                                                                                       0.05s
  ✓ admin can access projects.edit                                                                                                                      0.05s
  ✓ admin can update a project                                                                                                                          0.05s
  ✓ admin can access projects.delete                                                                                                                    0.05s
  ✓ admin can destroy a project                                                                                                                         0.05s
  ✓ unauthenticated user cannot access projects.index                                                                                                   0.04s
  ✓ unauthenticated user cannot access projects.create                                                                                                  0.05s
  ✓ unauthenticated user cannot store a project                                                                                                         0.04s
  ✓ unauthenticated user cannot access projects.edit                                                                                                    0.04s
  ✓ unauthenticated user cannot update a project                                                                                                        0.06s
  ✓ unauthenticated user cannot access projects.delete                                                                                                  0.05s
  ✓ unauthenticated user cannot destroy a project                                                                                                       0.04s
  ✓ project creation link is visible for users with create permission                                                                                   0.06s
  ✓ project edit link is visible for users with edit permission                                                                                         0.06s
  ✓ project delete link is visible for users with delete permission                                                                                     0.05s
  ✓ project delete link is not visible for users without delete permission

+++ app/Http/Controllers/Admin/ProjectController.php
@@ -14,6 +14,7 @@ class ProjectController extends \App\Http\Controllers\Controller
      */
     public function index()
     {
+        $this->authorize('index project');
         $projects = Project::all();
         return view('admin.projects.index', compact('projects'));
     }
@@ -23,6 +24,7 @@ class ProjectController extends \App\Http\Controllers\Controller
      */
     public function create()
     {
+        $this->authorize('create project');
         return view('admin.projects.create');
     }
 
@@ -31,6 +33,7 @@ class ProjectController extends \App\Http\Controllers\Controller
      */
     public function store(ProjectStoreRequest $request): \Illuminate\Http\RedirectResponse
     {
+        $this->authorize('create project');
         $validated = $request->validated();
 
         $project = new Project();
@@ -46,6 +49,7 @@ class ProjectController extends \App\Http\Controllers\Controller
      */
     public function show(Project $project)
     {
+        $this->authorize('show project');
         return view('admin.projects.show', compact('project'));
     }
 
@@ -54,6 +58,7 @@ class ProjectController extends \App\Http\Controllers\Controller
      */
     public function edit(Project $project)
     {
+        $this->authorize('edit project');
         return view('admin.projects.edit', ['project' => $project]);
     }
 
@@ -62,6 +67,7 @@ class ProjectController extends \App\Http\Controllers\Controller
      */
     public function update(ProjectUpdateRequest $request, Project $project)
     {
+        $this->authorize('edit project');
         $validated = $request->validated();
 
         $project->name = $validated['name'];
@@ -76,6 +82,7 @@ class ProjectController extends \App\Http\Controllers\Controller
      */
     public function delete(Project $project)
     {
+        $this->authorize('delete project');
         return view('admin.projects.delete', compact('project'));
     }
 
@@ -84,6 +91,7 @@ class ProjectController extends \App\Http\Controllers\Controller
      */
     public function destroy(Project $project)
     {
+        $this->authorize('delete project');
         $projectName = $project->name;
         $project->delete();
         return redirect()->route('projects.index')->with('status', "Project {$projectName} is verwijderd");
